Create README.md
In this data analysis project we performed a customer segmentation on a specific group of mall customers. We identified the best possible cluster using KMeans unsupervised machine learning algorithm to find the univariate, bivariate and multivariate clusters. Once we identified the clusters, summary statistics was performed on these to identify the best marketing group.